+++ src/api-demo-styles.ts
@@ -1,0 +1,104 @@
+import { css } from 'lit-element';
+
+export default css`
+  pre {
+    white-space: pre-wrap;
+  }
+
+  button {
+    position: absolute;
+    top: 0.5rem;
+    right: 0.5rem;
+    text-transform: uppercase;
+    border: none;
+    border-radius: 0.25em;
+    cursor: pointer;
+    background: var(--ave-button-background, rgba(0, 0, 0, 0.3));
+    color: var(--ave-button-color, #fff);
+  }
+
+  button:focus,
+  button:hover {
+    background: var(--ave-button-active-background, rgba(0, 0, 0, 0.6));
+  }
+
+  api-viewer-demo,
+  api-viewer-demo-layout {
+    display: block;
+  }
+
+  [part='demo-tabs'],
+  [part='demo-output'] {
+    border-top: solid 1px var(--ave-border-color);
+  }
+
+  [part='demo-tabs'] [part='tab-panel'] {
+    box-sizing: border-box;
+    position: relative;
+    background: #fafafa;
+  }
+
+  [part='demo-output'] {
+    padding: 1.5rem;
+    text-align: initial;
+    transform: translateZ(0);
+    overflow: hidden;
+  }
+
+  .source {
+    position: relative;
+  }
+
+  [part='knobs'] {
+    display: flex;
+    padding: 0 1rem 1rem;
+  }
+
+  [part='knobs-column'] {
+    width: 50%;
+  }
+
+  [part='knobs-header'] {
+    font-size: 1rem;
+    font-weight: bold;
+    margin: 1rem 0 0.25rem;
+  }
+
+  td {
+    padding: 0.25rem 0.25rem 0.25rem 0;
+    font-size: 0.9375rem;
+    white-space: nowrap;
+  }
+
+  [part='event-log'] {
+    display: block;
+    padding: 0 1rem;
+    min-height: 50px;
+    max-height: 200px;
+    overflow: auto;
+  }
+
+  [part='event-record'] {
+    margin: 0 0 0.25rem;
+    font-family: var(--ave-monospace-font);
+    font-size: 0.875rem;
+  }
+
+  [part='event-record']:first-of-type {
+    margin-top: 1rem;
+  }
+
+  [part='event-record']:last-of-type {
+    margin-bottom: 1rem;
+  }
+
+  @media (max-width: 480px) {
+    [part='knobs'] {
+      flex-direction: column;
+    }
+
+    [part='knobs-column']:not(:last-child) {
+      margin-bottom: 1rem;
+    }
+  }
+`;