+++ src/api-viewer-demo-snippet.ts
@@ -11,14 +11,15 @@ import { htmlRender } from 'highlight-ts/es/render/html';
 import { registerLanguages } from 'highlight-ts/es/languages';
 import { XML } from 'highlight-ts/es/languages/xml';
 import { init, process } from 'highlight-ts/es/process';
-import {
-  CSSPropertyInfo,
-  KnobValues,
-  KnobValue,
-  SlotValue
-} from './lib/types.js';
+import { CSSPropertyInfo, KnobValues, SlotValue } from './lib/types.js';
 import { CSS } from './lib/highlight-css.js';
-import { getSlotTemplate, normalizeType } from './lib/utils.js';
+import {
+  getTemplate,
+  getTemplateNode,
+  isTemplate,
+  normalizeType,
+  TemplateTypes
+} from './lib/utils.js';
 import highlightTheme from './lib/highlight-theme.js';
 
 // register languages
@@ -29,9 +30,11 @@ const highlighter = init(htmlRender, {
   classPrefix: ''
 });
 
+const { PREFIX, SLOT, SUFFIX, WRAPPER } = TemplateTypes;
+
 const INDENT = '  ';
 
-const unindent = (text: string) => {
+const unindent = (text: string, prepend: string) => {
   if (!text) return text;
   const lines = text.replace(/\t/g, INDENT).split('\n');
   const indent = lines.reduce((prev: number | null, line: string) => {
@@ -43,50 +46,93 @@ const unindent = (text: string) => {
     return (lineIndent as number) < prev ? lineIndent : prev;
   }, null);
 
-  return lines.map(l => INDENT + l.substr(indent as number)).join('\n');
+  return lines.map(l => prepend + l.substr(indent as number)).join('\n');
+};
+
+const getTplContent = (
+  template: HTMLTemplateElement,
+  prepend: string
+): string => {
+  const tpl = template.innerHTML.replace(/\s+$/, '').replace(/(="")/g, '');
+  return unindent(tpl, prepend);
 };
 
 const renderSnippet = (
+  id: number,
   tag: string,
   values: KnobValues,
   slots: SlotValue[],
   cssProps: CSSPropertyInfo[]
 ): TemplateResult => {
-  let markup = `<${tag}`;
+  let markup = '';
+  const prefix = getTemplate(id, tag, PREFIX);
+  if (isTemplate(prefix)) {
+    markup += `${getTplContent(prefix, '').trim()}\n`;
+  }
+
+  let prepend = '';
+  let wrap = null;
+
+  const wrapper = getTemplate(id, tag, WRAPPER);
+  const wrapNode = getTemplateNode(wrapper);
+  if (wrapNode) {
+    prepend = INDENT;
+    const match = wrapNode.outerHTML.match(/<([a-z]+)[^>]*(?<!\/)>/);
+    if (match) {
+      wrap = wrapNode.tagName.toLowerCase();
+      markup += `${match[0]}\n${INDENT}`;
+    }
+  }
+
+  markup += `<${tag}`;
   Object.keys(values)
     .sort((a, b) => (a > b ? 1 : -1))
     .forEach((key: string) => {
-      const knob: KnobValue = values[key];
-      const attr = knob.attribute || key;
-      switch (normalizeType(knob.type)) {
+      const { value, type, attribute } = values[key];
+      const attr = attribute || key;
+      switch (normalizeType(type)) {
         case 'boolean':
-          markup += knob.value ? ` ${attr}` : '';
+          markup += value ? ` ${attr}` : '';
+          break;
+        case 'select':
+          markup += value !== '' ? ` ${attr}="${value}"` : '';
           break;
         default:
-          markup += knob.value != null ? ` ${attr}="${knob.value}"` : '';
+          markup += value != null ? ` ${attr}="${value}"` : '';
           break;
       }
     });
 
   markup += `>`;
 
-  const template = getSlotTemplate(tag);
-  if (template instanceof HTMLTemplateElement) {
-    const tpl = template.innerHTML.replace(/\s+$/, '').replace(/(="")/g, '');
-    markup += unindent(tpl);
-    markup += `\n`;
+  const template = getTemplate(id, tag, SLOT);
+  if (isTemplate(template)) {
+    markup += `${getTplContent(template, `${prepend}${INDENT}`)}\n${prepend}`;
   } else if (slots.length) {
-    slots.forEach(slot => {
-      const { name, content } = slot;
-      const div = name ? `<div slot="${name}">` : '<div>';
-      markup += `\n${INDENT}${div}${content}</div>`;
-    });
-    markup += `\n`;
+    if (slots.length === 1 && !slots[0].name) {
+      markup += slots[0].content;
+    } else {
+      markup += slots.reduce((result: string, slot) => {
+        const { name, content } = slot;
+        const line = name ? `<div slot="${name}">${content}</div>` : content;
+        return `${result}\n${prepend}${INDENT}${line}`;
+      }, '');
+      markup += `\n${prepend}`;
+    }
   }
 
   markup += `</${tag}>`;
 
-  const cssValues = cssProps.filter(p => p.value !== p.defaultValue);
+  if (wrap) {
+    markup += `\n</${wrap}>`;
+  }
+
+  const suffix = getTemplate(id, tag, SUFFIX);
+  if (isTemplate(suffix)) {
+    markup += `\n${getTplContent(suffix, '').trim()}\n`;
+  }
+
+  const cssValues = cssProps.filter(p => p.value !== p.default);
   if (cssValues.length) {
     markup += `\n<style>\n${INDENT}${tag} {\n`;
     cssValues.forEach(prop => {
@@ -108,14 +154,16 @@ const renderSnippet = (
 export class ApiViewerDemoSnippet extends LitElement {
   @property({ type: String }) tag = '';
 
-  @property({ attribute: false, hasChanged: () => true })
+  @property({ attribute: false })
   knobs: KnobValues = {};
 
-  @property({ attribute: false, hasChanged: () => true })
+  @property({ attribute: false })
   slots: SlotValue[] = [];
 
-  @property({ attribute: false, hasChanged: () => true })
+  @property({ attribute: false })
   cssProps: CSSPropertyInfo[] = [];
+
+  @property({ type: Number }) vid?: number;
 
   static get styles() {
     return [
@@ -131,7 +179,13 @@ export class ApiViewerDemoSnippet extends LitElement {
 
   protected render() {
     return html`
-      ${renderSnippet(this.tag, this.knobs, this.slots, this.cssProps)}
+      ${renderSnippet(
+        this.vid as number,
+        this.tag,
+        this.knobs,
+        this.slots,
+        this.cssProps
+      )}
     `;
   }