+++ src/lib/markdown.ts
@@ -4,6 +4,8 @@ import { unsafeHTML } from 'lit-html/directives/unsafe-html.js';
 import marked from 'marked/lib/marked.esm.js';
 import DOMPurify from 'dompurify';
 
+marked.setOptions({ headerIds: false });
+
 export const parse = (markdown?: string): TemplateResult => {
   if (!markdown) {
     return html`
@@ -12,6 +14,11 @@ export const parse = (markdown?: string): TemplateResult => {
   }
 
   return html`
-    ${unsafeHTML(DOMPurify.sanitize(marked(markdown)))}
+    ${unsafeHTML(
+      DOMPurify.sanitize(marked(markdown)).replace(
+        /<(h[1-6]|a|p|ul|ol|li|pre|code|strong|em|blockquote|del)(\s+href="[^"]+")*>/g,
+        '<$1 part="md-$1"$2>'
+      )
+    )}
   `;
 };